refactor(client): :zap: added resizable feature and modularized bento grid side
