+++ packages/client/package.json
@@ -34,6 +34,7 @@
     "react-day-picker": "8.10.1",
     "react-dom": "19.0.0-rc-66855b96-20241106",
     "react-hook-form": "^7.53.2",
+    "react-resizable-panels": "^2.1.7",
     "reactflow": "^11.11.4",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
